math.h fixed

+++ src/getBinary.c
@@ -1,5 +1,5 @@
 #include "headers.h"
-#include <Math.h>
+#include <math.h>
 
 /*
 @brief This function converts a decimal number to its binary representation
@@ -12,7 +12,7 @@ significant bit stored last and least significant stored first)
 */
 
 int* decToBinary(int number, int* length) {
-  *length = log10((double)number)/log10((double)2);
+  (double) *length = log10((double)number)/log10((double)2);
   int* binary = malloc(sizeof(int) * *length);
   int i = 0;
   while (number > 0) {
